Update move_range_preview.js
Goal is to create an overlay that shows proximity to given tiles, i.e. how much control player 1 or 2 has over a given area. 
An altered version of getPositionsInMoveRange has been created that uncaps the search. This is then translated to # of turns to reach the given tile. Then, those functions can be used to identify the default weak/strong sides of the map, or potentially map control from a given board state if filters for unit types can be applied.